requirements.txt
discord-py
python-dotenv
asyncio
steam